# agri-frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "AgriSmart Benin API",
  "description": "Backend Monolithique Modulaire pour le Challenge Agriculture Intelligente",
  "version": "4.0.0"
 },
 "paths": {
...
  "/api/v1/storage/lots": {
   "post": {
    "tags": [
     "Conseiller de stockage"
    ],
    "summary": "Déclarer un stock",
    "operationId": "create_lot_api_v1_storage_lots_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/LotCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/storage/lots/me": {
   "get": {
    "tags": [
     "Conseiller de stockage"
    ],
    "summary": "Mes stocks, avec le risque et les contrôles à faire",
    "operationId": "my_lots_api_v1_storage_lots_me_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/storage/lots/{lot_id}/checks": {
   "post": {
    "tags": [
     "Conseiller de stockage"
    ],
    "summary": "Enregistrer un contrôle du stock",
    "operationId": "add_check_api_v1_storage_lots__lot_id__checks_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "lot_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Lot Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CheckCreate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/storage/lots/{lot_id}/status": {
   "patch": {
    "tags": [
     "Conseiller de stockage"
    ],
    "summary": "Clore un stock (vendu, consommé, perdu)",
    "operationId": "close_lot_api_v1_storage_lots__lot_id__status_patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "lot_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Lot Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/LotStatus"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/storage/lots/{lot_id}/advice": {
   "post": {
    "tags": [
     "Conseiller de stockage"
    ],
    "summary": "Conseil personnalisé : conservation, vendre ou stocker (IA)",
    "operationId": "advice_api_v1_storage_lots__lot_id__advice_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "lot_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Lot Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/storage/overview": {
   "get": {
    "tags": [
     "Conseiller de stockage"
    ],
    "summary": "Stocks et pertes déclarés (agents)",
    "operationId": "overview_api_v1_storage_overview_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "CheckCreate": {
    "properties": {
     "moisture_pct": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 40.0,
        "minimum": 3.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Moisture Pct"
     },
     "insects_seen": {
      "type": "boolean",
      "title": "Insects Seen",
      "default": false
     },
     "mold_seen": {
      "type": "boolean",
      "title": "Mold Seen",
      "default": false
     },
     "quantity_kg": {
      "anyOf": [
       {
        "type": "number",
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Quantity Kg",
      "description": "Quantité restante constatée"
     },
     "note": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 500
       },
       {
        "type": "null"
       }
      ],
      "title": "Note"
     }
    },
    "type": "object",
    "title": "CheckCreate"
   },
...
   "Department": {
    "type": "string",
    "enum": [
     "Alibori",
     "Atacora",
     "Atlantique",
     "Borgou",
     "Collines",
     "Couffo",
     "Donga",
     "Littoral",
     "Mono",
     "Ouémé",
     "Plateau",
     "Zou"
    ],
    "title": "Department"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "LotCreate": {
    "properties": {
     "product": {
      "type": "string",
      "maxLength": 60,
      "minLength": 2,
      "title": "Product",
      "examples": [
       "Maïs"
      ]
     },
     "quantity_kg": {
      "type": "number",
      "exclusiveMinimum": 0.0,
      "title": "Quantity Kg"
     },
     "harvest_date": {
      "type": "string",
      "format": "date",
      "title": "Harvest Date"
     },
     "storage_method": {
      "type": "string",
      "enum": [
       "sac_hermetique",
       "sac_polypropylene",
       "grenier_traditionnel",
       "magasin_ventile",
       "fut_hermetique",
       "autre"
      ],
      "title": "Storage Method"
     },
     "moisture_pct": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 40.0,
        "minimum": 3.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Moisture Pct",
      "description": "Humidité mesurée (humidimètre)"
     },
     "treatment": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200
       },
       {
        "type": "null"
       }
      ],
      "title": "Treatment",
      "description": "Produit de conservation utilisé (homologué)"
     },
     "land_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Land Id"
     },
     "department": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Department"
       },
       {
        "type": "null"
       }
      ]
     },
     "commune": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Commune"
     },
     "client_ref": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^[A-Za-z0-9_-]{8,64}$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Client Ref",
      "description": "Identifiant unique généré par l'application (ex. UUID). Un envoi rejoué avec le même client_ref (file d'attente hors ligne) renvoie la ressource existante au lieu d'un doublon."
     }
    },
    "type": "object",
    "required": [
     "product",
     "quantity_kg",
     "harvest_date",
     "storage_method"
    ],
    "title": "LotCreate"
   },
   "LotStatus": {
    "properties": {
     "status": {
      "type": "string",
      "enum": [
       "vendu",
       "consomme",
       "perdu",
       "transfere"
      ],
      "title": "Status"
     },
     "quantity_lost_kg": {
      "anyOf": [
       {
        "type": "number",
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Quantity Lost Kg"
     }
    },
    "type": "object",
    "required": [
     "status"
    ],
    "title": "LotStatus"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}